Extract CustomNamesContext and CustomNameField into components, following naming style of react-admin useRecordContext, etc.
