+++ Development/src/components/CustomNamesContext.js
@@ -1,0 +1,4 @@
+import { createContext } from 'react';
+
+export const CustomNamesContext = createContext();
+export default CustomNamesContext;
