+++ Development/src/components/CustomNamesContextProvider.js
@@ -1,0 +1,7 @@
+import React from 'react';
+import CustomNamesContext from './CustomNamesContext';
+
+export const CustomNamesContextProvider = props => (
+    <CustomNamesContext.Provider {...props} />
+);
+export default CustomNamesContextProvider;
