+++ Development/src/components/useCustomNamesContext.js
@@ -1,0 +1,5 @@
+import { useContext } from 'react';
+import CustomNamesContext from './CustomNamesContext';
+
+export const useCustomNamesContext = () => useContext(CustomNamesContext);
+export default useCustomNamesContext;
